* commit 'ca8563129858592d92e3f69ed830febb362c3c88':
  Fix `-Wnarrowing`

+++ include/pacbio/pancake/Minimizers.h
@@ -121,8 +121,13 @@ bool CollectSeedHits(std::vector<SeedHit>& hits, const PacBio::Pancake::SeedDB::
                     queryPos = queryLen - (decodedQuery.pos + querySpan);
                 }
 
-                SeedHit hit{decodedTarget.seqID, isRev,     targetPos, queryPos,
-                            targetSpan,          querySpan, 0};
+                SeedHit hit{static_cast<int32_t>(decodedTarget.seqID),
+                            isRev,
+                            targetPos,
+                            queryPos,
+                            targetSpan,
+                            querySpan,
+                            0};
 
                 hits.emplace_back(hit);
             }
